Address some compilation issues and complete ss7_event2str
There were a few variables that were set but unused that were causing
compilation errors due to the compilation flags used (-Werror).  These
are now corrected.

This patch also finishes out the functionality of ss7_event2str so that
it returns the correct name for all event types it knows about.  This
change is related to the Asterisk 11 Who Hung Up? work.


git-svn-id: https://origsvn.digium.com/svn/libss7/branches/1.0@297 284bc105-fd19-0410-aa88-94cd1a20766c